[refactor] Score파일 함수 세분화 #775
