+++ components/modal/afterGame/CheckedScore.tsx
@@ -1,0 +1,16 @@
+import { TeamScore } from 'types/scoreTypes';
+import styles from 'styles/modal/afterGame/AfterGameModal.module.scss';
+
+interface CheckedScoreProps {
+  result: TeamScore;
+}
+
+export default function CheckedScore({ result }: CheckedScoreProps) {
+  return (
+    <div className={styles.finalScore}>
+      <div>{result.myTeamScore}</div>
+      <div>:</div>
+      <div>{result.enemyTeamScore}</div>
+    </div>
+  );
+}
